+++ x/foundation/keeper/keys.go
@@ -22,8 +22,6 @@ var (
 	grantKeyPrefix = []byte{0x20}
 
 	poolKey = []byte{0x30}
-
-	govMintKey = []byte{0x40}
 )
 
 // must be constant